% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcLibrary.R
\name{vcLibrary}
\alias{vcLibrary}
\title{Loads a specific version of a package}
\usage{
vcLibrary(package, version, ...)
}
\arguments{
\item{package}{character vector of the names of packages which should be loaded.}

\item{version}{character vector of the required versions of packages to be loaded.}

\item{...}{arguments to to be passed to library}
}
\value{
invisible NULL
}
\description{
Loads a specific version of a package
}
\details{
The function will attempt to load the specific version(s) of the requested package(s) into your namespace.
To support multiple versions of packages being installed side-by-side, new directories were created near your primary library based on package name and version during installation.  This supports reusing the same version of a package inside multiple scripts or projects.  The function uses the knowledge of how this was performed to use the correct library to ensure relevant dependencies are also installed.
The version matching includes the use of 'x' for variable versioning.  Package versions are often of the sort major.minor.patch - if your code is resilient against patch versions of the code you could request version 1.1.x which would install the most recent version which has a major version of 1, a minor version of 1 and any patch version.  Note, 1.x.1 is not a supported form of version matching - this will be interpreted as 1.x
}
\examples{
vcLibrary('coin','1.1x')
# Identical result as
vcLibrary('coin','1.1-3')
}
